add test visualization
