+++ web/app/index.html
@@ -13,8 +13,8 @@
   <header>
     <nav>
       <a href="/">RJ45</a>
-      <a href="/tests.html">Tests</a>
-      <a href="/login.html">Iniciar sesión</a>
+      <a href="/tests">Tests</a>
+      <a href="/login">Iniciar sesión</a>
     </nav>
   </header>
   <div class="container">
