+++ web/app/tests.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Inicio - RJ45</title>
+  <title>Tests - RJ45</title>
   <link rel="stylesheet" href="css/globals.css">
   <link rel="stylesheet" href="css/tests.css">
   <script type="module" src="js/tests.js"></script>
@@ -14,8 +14,8 @@
   <header>
     <nav>
       <a href="/">RJ45</a>
-      <a href="/tests.html">Tests</a>
-      <a href="/login.html">Iniciar sesión</a>
+      <a href="/tests">Tests</a>
+      <a href="/login">Iniciar sesión</a>
     </nav>
   </header>
   <div class="container">
@@ -23,7 +23,7 @@
     <div class="actions">
       <input type="text" id="search-bar" placeholder="Search tests...">
       <button id="search-btn">Search</button>
-      <button id="add-test-btn">Add New Test</button>
+      <button id="add-test-btn">+</button>
     </div>
     <table id="tests-table">
       <thead>
